More flexible cross comparisons between sources
Developed an approach for templating queries that allows for various core types of data element and ones specific for each source.

A dict with a list of tuples for each source is created to fill out the comparison part of the categories query.

The data elements are strictly name spaced into groups which has necessitated changes in naming in the creation of the truthtables.