Added ORDER BY source_id to remove stochastic effects
Previously when selecting a single source_id from mag or openalex where more than one was assocaited with the same DOI there was a stochastic effect where multiple PaperIds could have the same citation count. Because PaperId is a primary key for this source table and they are ordered in order of addition sorting on PaperId DESC is guaranteed to give the most recently added object that has the most citations.